Finished splitting out transactions.rs from root into submodules.
