--- contracts/mixnet/src/transactions.rs
@@ -1,10 +1,0 @@
-// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
-
-pub(crate) const OLD_DELEGATIONS_CHUNK_SIZE: usize = 500;
-
-// approximately 1 day (assuming 5s per block)
-pub(crate) const MINIMUM_BLOCK_AGE_FOR_REWARDING: u64 = 17280;
-
-// approximately 30min (assuming 5s per block)
-pub(crate) const MAX_REWARDING_DURATION_IN_BLOCKS: u64 = 360;
